requirements.txt
django
djangorestframework
requests
django-environ
requests-mock